all_requirements.txt
absl-py==0.7.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-project==0.8.2
asn1crypto==0.24.0
astor==0.7.1
astroid==2.2.5
astropy==3.1.2
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==3.1.0
blessings==1.7
bokeh==1.0.4
boto==2.49.0
Bottleneck==1.2.1
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==0.8.0
clyent==1.2.2
colorama==0.4.1
comtypes==1.1.7
contextlib2==0.5.5
cryptography==2.6.1
curses==2.2+utf8
curses-menu==0.5.0
cycler==0.10.0
Cython==0.29.6
cytoolz==0.9.0.1
dask==1.1.4
datashape==0.5.4
decorator==4.3.2
defusedxml==0.5.0
distributed==1.26.0
docopt==0.6.2
docutils==0.14
earthengine-api==0.1.173
ee==0.2
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.10
Flask==1.0.2
Flask-Cors==3.0.7
gast==0.2.2
gee2drive==0.0.9
gevent==1.4.0
gitdb2==2.0.5
GitPython==2.1.11
glob2==0.6
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
greenlet==0.4.15
grpcio==1.16.1
h5py==2.8.0
heapdict==1.0.0
html5lib==1.0.1
httplib2==0.12.1
idna==2.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.0.0
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.15
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.3
Jinja2==2.10
joblib==0.13.2
jsonpickle==1.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.3
jupyterlab-server==0.2.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==18.0.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
libarchive-c==2.8
llvmlite==0.28.0
locket==0.2.0
lxml==4.3.0
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
mock==2.0.0
more-itertools==6.0.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
nltk==3.4
nose==1.3.7
notebook==5.7.8
numba==0.43.0
numexpr==2.6.9
numpy==1.16.2
numpydoc==0.8.0
odo==0.5.1
olefile==0.46
opencv-python==4.0.1
openpyxl==2.6.1
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
partd==0.3.9
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.1
pbr==5.1.3
pendulum==2.0.4
pep8==1.7.1
pickleshare==0.7.5
Pillow==5.4.1
pipreqs==0.4.9
pkginfo==1.5.0.1
plotly==3.10.0rc1
pluggy==0.9.0
ply==3.11
progressbar==2.5
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.6.1
psutil==5.6.1
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
pyflakes==2.1.1
Pygments==2.3.1
pylint==2.3.1
pyodbc==4.0.26
pyOpenSSL==19.0.0
pyparsing==2.3.1
pyrsistent==0.14.11
PySocks==1.6.8
pytest==4.3.0
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-dateutil==2.8.0
pytz==2018.9
pytzdata==2018.9
PyWavelets==1.0.2
pywin32==223
pywinpty==0.5.5
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.5.7
qtconsole==4.4.4
QtPy==1.6.0
requests==2.21.0
retrying==1.3.3
rope==0.12.0
rsa==4.0
ruamel-yaml==0.15.46
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
Shapely==1.6.4.post1
simplegeneric==0.8.1
singledispatch==3.4.0.3
sip==4.19.13
six==1.12.0
smmap2==2.0.5
snowballstemmer==1.2.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.8
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.0
spyder==3.3.3
spyder-kernels==0.4.2
SQLAlchemy==1.3.1
statsmodels==0.9.0
sympy==1.3
tables==3.4.4
tblib==1.3.2
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
torch==0.3.1.post2
tornado==6.0.1
tqdm==4.31.1
traitlets==4.3.2
typed-ast==1.3.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.1
xlrd==1.2.0
XlsxWriter==1.1.5
xlwings==0.15.3
xlwt==1.3.0
yarg==0.1.9
zict==0.1.4
zipp==0.3.3
